[CHORE] Delete middleware directory and move the file
